Create functionality to add categories
Items now have a category to sort them by. Users can see all items by
clicking on home or only view items within a certain category by
clicking on it's name in the sidebar. If a user enters a new category
when making a new item, that new category is created and added to the
sidebar.